+++ include/materials.hpp
@@ -6,10 +6,10 @@
 class Material {
     public:
         Color color;
-        float ambient;
-        float diffuse;
-        float specular;
-        float shininess;
+        double ambient;
+        double diffuse;
+        double specular;
+        double shininess;
 
         Material();
 };
